refactor: extract loginPage of homePage
