+++ src/Router.tsx
@@ -1,6 +1,7 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 
 import HomePage from "pages/HomePage";
+import LoginPage from "pages/LoginPage";
 
 
 function Router() {
@@ -8,6 +9,7 @@ function Router() {
         <BrowserRouter>
             <Routes>
                 <Route path="/" element={<HomePage />} />
+                <Route path="login" element={<LoginPage />} />
             </Routes>
         </BrowserRouter>
     );
