+++ src/components/Login/index.tsx
@@ -4,14 +4,17 @@ import { ReactComponent as EmailIco } from 'assets/icons/email.svg';
 import { ReactComponent as PassIco } from 'assets/icons/pass.svg';
 
 import { LoginButtonNormal, LoginButtonWarm, LoginButtons, LoginContainer, LoginField, LoginTitle } from "./stlyle";
+import { useNavigate } from "react-router";
 
 
 function Login() {
     const [userName, setUserName] = useState("");
 
+    const navigate = useNavigate();
+
     function login() {
-        localStorage.setItem('token', userName)
-        window.location.reload()
+        localStorage.setItem('token', userName);
+        navigate('/');
     }
 
     return (
@@ -29,8 +32,8 @@ function Login() {
             </LoginField>
             <LoginButtons>
                 <LoginButtonNormal onClick={login} >Entrar</LoginButtonNormal>
-                <LoginButtonNormal onClick={()=> alert('Serviço não implementado')} >Cadastro</LoginButtonNormal>
-                <LoginButtonWarm onClick={()=> alert('Serviço não implementado')} >Recuperar senha</LoginButtonWarm>
+                <LoginButtonNormal onClick={() => alert('Serviço não implementado')} >Cadastro</LoginButtonNormal>
+                <LoginButtonWarm onClick={() => alert('Serviço não implementado')} >Recuperar senha</LoginButtonWarm>
             </LoginButtons>
         </LoginContainer>
     );
